Draft of Genre Suggester.
Created a basic yes or no expert system to work off of, based on an example provided by Buckinghamshire New University. Program suggests a music genre based on a series of yes or no questions and checking the stored values against set metrics before making the suggestion. This will serve as a basis and we can work further upon that. Hopefully a system where a user can type a word as an answer as opposed to yes or no, as well as additional rules to meet the coursework requirements.