Attempted Song suggestion extention
Added means to suggest songs based on the user's favourite/ highest scoring genre. However, instead of seeing if the Genre is equal to a certain value (i.e. Genre = metal or Genre = punk), it only picks metal as that is the first option and sets Genre's value to metal. Need to figure out how to make that value flexible to whatever scores the highest.